fix build, remove unnecessary types
