--- .babelrc
@@ -1,4 +1,0 @@
-{
-    "presets": [["@babel/preset-react"], ["@babel/preset-typescript"]],
-    "plugins": []
-  }
