+++ src/aws-widgets/aws-ecs-deployments.tsx
@@ -26,11 +26,42 @@ import {
 import { getAwsCredentialsProvider } from '../utils/utils.js';
 import TaskDefinitionBody from '../components/task-definition-body.js';
 import DeploymentRow from '../components/deployment-row.js';
-import {
-  AwsEcsDeployments as AwsEcsDeploymentsType,
-  Deployment,
-  Task
-} from '../ops-types.js';
+import { AwsEcsDeployments as AwsEcsDeploymentsProps } from '../ops-types.js';
+import { Image } from '../utils/aws-ecs-utils.js';
+
+type Task = {
+  taskId?: string;
+  startTime?: Date;
+  stopTime?: Date;
+  status?: string;
+  group?: string;
+  version?: number;
+  cwLogsArn?: string;
+};
+
+export type TaskDefinition = {
+  taskDefinitionArn?: string;
+  cpu?: string;
+  memory?: string;
+  roleArn?: string;
+  execRoleArn?: string;
+  containers?: Image[];
+  tasks?: Task[];
+};
+
+export type Deployment = {
+  deploymentId?: string;
+  status?: string;
+  startTime?: Date;
+  runningCount?: number;
+  pendingCount?: number;
+  desiredCount?: number;
+  taskDefinition?: TaskDefinition;
+};
+
+type AwsEcsDeploymentsType = AwsEcsDeploymentsProps & {
+  deployments?: Deployment[];
+};
 
 type AwsEcsDeploymentsOverrides = {
   stoppedTaskId?: string;
