+++ src/aws-widgets/aws-json-tree.tsx
@@ -1,7 +1,8 @@
+import React from 'react';
+import isEmpty from 'lodash.isempty';
 import { BaseProvider, BaseWidget, TinyStacksError } from '@tinystacks/ops-core';
 import CloudControl, { ResourceDescription } from 'aws-sdk/clients/cloudcontrol';
 import { AwsSdkVersionEnum } from '../aws-provider/aws-credentials/aws-credentials-type.js';
-import isEmpty from 'lodash.isempty';
 import { getAwsCredentialsProvider } from '../utils/utils.js';
 import { AwsJsonTree as AwsJsonTreeProps } from '../ops-types.js';
