+++ src/components/ecs-env-vars-modal.tsx
@@ -1,8 +1,8 @@
+import React from 'react';
 import {
   Button, Modal, ModalBody, ModalCloseButton, ModalContent, ModalFooter, ModalHeader, ModalOverlay, Stack, Table, Tbody, Td, Th, Thead, Tr, useDisclosure
 } from '@chakra-ui/react';
-import React from 'react';
-import { Image } from '../ops-types.js';
+import { Image } from '../utils/aws-ecs-utils.js';
 
 export default function EcsEnvVarsModal (props: { image: Image }) {
   const { isOpen, onOpen, onClose } = useDisclosure();
